taint/test/mochitest/test_local_storage_sinks.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset="utf-8">
    <title>Test LocalStorage Sinks</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
    <script>

      let string_content = "hello";
      let sink_names = [
	  "localStorage.setItem",
	  "localStorage.setItem(key)"
      ];

      let i = 0;
      
      SimpleTest.waitForExplicitFinish();
      addEventListener("__taintreport", (report) => {
          is(report.detail.str, string_content, "Check sink string content");

          let flow = report.detail.str.taint[0].flow;
          is(flow[2].operation, sink_names[i]);

	  i += 1;
	  if (i >= sink_names.length) {
              SimpleTest.finish();
	  }
      }, false);

      function startTest() {
	  let taint_string = String.tainted(string_content);
	  localStorage.setItem("tainted", taint_string);
	  localStorage.setItem(taint_string, "tainted");
      }
      
    </script>
  </head>

  <body onload="startTest();">

</html>
